Add files via upload
We got this code to work on Port D, where it simply assigns the different power levels to 3 separate pins. However, this code is altered to accommodate for Port F, which is not accessible with the prototype board. It does not have the interrupts, and should eventually be added to the timer counter initialization code.